[feat] extracting projects models to new files
