+++ src/app/shared/models/project-item.model.ts
@@ -1,0 +1,5 @@
+export interface ProjectItem {
+    id: string,
+    title: string,
+    url: string
+}
