+++ src/app/shared/models/projects-card-details.model.ts
@@ -1,0 +1,6 @@
+import { projectCardInfo } from "src/app/modules/projects/models/project-card-info.model";
+
+export interface ProjectsCardDetails {
+    currentProjects: projectCardInfo[],
+    finishedProjects: projectCardInfo[]
+}
